requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
gensim==4.0.1
streamlit==0.85.0
pandas==1.3.1
numpy==1.21.1
nltk==3.6.2
sklearn==0.0
lyrics-extractor==3.0.1
spotipy==2.18.0
python-Levenshtein==0.12.2
mysql-connector-python==8.0.06